fetch missions data from the API
