+++ src/components/Missions/Mission.js
@@ -5,13 +5,13 @@ import './Mission.css';
 const Mission = ({ mission }) => (
   <tr>
     <td>
-      <p className="mission-name-content">{mission.name}</p>
+      <p className="mission-name-content">{mission.mission_name}</p>
     </td>
     <td>
       <p className="desc-content">{mission.description}</p>
     </td>
     <td className="status-mission">
-      <p className="status-content">{mission.member ? 'Active Member' : 'NOT A MEMBER'}</p>
+      <p className="status-content">NOT A MEMBER</p>
     </td>
     <td className="join-mission">
       <button type="button" className="join-mission-btn">
@@ -23,9 +23,9 @@ const Mission = ({ mission }) => (
 
 Mission.propTypes = {
   mission: PropTypes.shape({
-    name: PropTypes.string.isRequired,
+    mission_name: PropTypes.string.isRequired,
     description: PropTypes.string.isRequired,
-    member: PropTypes.bool.isRequired,
+    member: PropTypes.bool,
   }).isRequired,
 };
