--- src/components/Missions/generateId.js
@@ -1,8 +1,0 @@
-// import { useId } from 'react-id-generator';
-
-function generate() {
-  const randomNumber = Math.random() * 1000;
-  return randomNumber;
-}
-
-export default generate;
